finish project

+++ src/features/comment/CommentCard.js
@@ -1,9 +1,12 @@
 import {
   Avatar,
+  Button,
+  Card,
   IconButton,
   Menu,
   MenuItem,
   Paper,
+  Popover,
   Stack,
   Typography,
 } from "@mui/material";
@@ -18,6 +21,7 @@ import useAuth from "../../hooks/useAuth";
 
 function CommentCard({ comment }) {
   const [anchorEl, setAnchorEl] = useState(null);
+  const [isConfirm, setIsConfirm] = useState(null);
   const open = Boolean(anchorEl);
   const dispatch = useDispatch();
   const commentId = comment._id;
@@ -32,10 +36,70 @@ function CommentCard({ comment }) {
     setAnchorEl(null);
   };
 
-  const handleDeleteComment = () => {
-    setAnchorEl(null);
-    dispatch(deleteComment({ commentId }));
+  const handleConfirmClose = () => {
+    setIsConfirm(null);
   };
+
+  const handleDeleteComment = (e) => {
+    setIsConfirm(e.currentTarget);
+    setAnchorEl(null);
+  };
+
+  const handleConfirmDelete = (e) => {
+    if (e === "YES") {
+      dispatch(deleteComment({ commentId }));
+    } else if (e === "CANCEL") {
+      setIsConfirm(null);
+      return;
+    }
+  };
+
+  const renderConfirmDelete = (
+    <Popover
+      id={comment._id}
+      open={Boolean(isConfirm)}
+      anchorEl={isConfirm}
+      onClose={handleConfirmClose}
+      anchorOrigin={{
+        vertical: "top",
+        horizontal: "left",
+      }}
+      transformOrigin={{
+        vertical: "center",
+        horizontal: "right",
+      }}
+      sx={{
+        borderRadius: 8,
+      }}
+    >
+      <Card
+        sx={{
+          width: 300,
+          height: 100,
+          p: 1,
+          alignContent: "center",
+          justifyContent: "center",
+          display: "flex",
+          flexDirection: "column",
+          backgroundColor: "#f5f3f0",
+          borderRadius: 0,
+        }}
+      >
+        <Typography
+          variant="subtitle2"
+          mt={1}
+          justifyContent="center"
+          display="flex"
+        >
+          Are you sure to delete this comment?
+        </Typography>
+        <Box sx={{ display: "flex", justifyContent: "center", p: 1 }}>
+          <Button onClick={() => handleConfirmDelete("YES")}>Yes</Button>
+          <Button onClick={() => handleConfirmDelete("CANCEL")}>Cancel</Button>
+        </Box>
+      </Card>
+    </Popover>
+  );
 
   const renderMenu = (
     <Menu
@@ -56,6 +120,7 @@ function CommentCard({ comment }) {
       <MenuItem sx={{ my: 1 }} onClick={handleDeleteComment}>
         Delete Comment
       </MenuItem>
+      {renderConfirmDelete}
     </Menu>
   );
   return (
